- rearrange source according to specs
